+++ R/propeller.R
@@ -19,12 +19,14 @@
 #' The \code{propeller} function calculates cell type proportions for each
 #' biological replicate, performs a variance stabilising transformation on the
 #' matrix of proportions and fits a linear model for each cell type or cluster
-#' using the \code{limma} framework. Propeller tests whether there is a
-#' difference in the cell type proportions between multiple groups. If there
-#' are only 2 groups, a t-test is used to calculate p-values, and if there are
-#' more than 2 groups, an F-test (ANOVA) is used. Cell type proportions of 1 or
-#' 0 are accommodated. Benjamini and Hochberg false discovery rates are
-#' calculated to account to multiple testing of cell types/clusters.
+#' using the \code{limma} framework. There are two options for the 
+#' transformation: arcsin square root or logit. Propeller tests whether there 
+#' is a difference in the cell type proportions between multiple groups. 
+#' If there are only 2 groups, a t-test is used to calculate p-values, and if 
+#' there are more than 2 groups, an F-test (ANOVA) is used. Cell type 
+#' proportions of 1 or 0 are accommodated. Benjamini and Hochberg false 
+#' discovery rates are calculated to account to multiple testing of 
+#' cell types/clusters.
 #'
 #' @aliases propeller
 #' @param x object of class \code{SingleCellExperiment} or \code{Seurat}
@@ -44,6 +46,9 @@
 #' proportions
 #' @param robust logical, if true performs robust empirical Bayes shrinkage of
 #' the variances
+#' @param transform a character scalar specifying which transformation of the 
+#' proportions to perform. Possible values include "asin" or "logit". Defaults
+#' to "asin".
 #'
 #' @return produces a dataframe of results
 #'
@@ -104,10 +109,10 @@
 #'   grp <- rep(c("grp1","grp2"),c(sum(numcells[1:2]),sum(numcells[3:4])))
 #'
 #'   propeller(clusters = clust, sample = biorep, group = grp,
-#'   robust = FALSE, trend = FALSE)
+#'   robust = FALSE, trend = FALSE, transform="asin")
 #'
 propeller <- function(x=NULL, clusters=NULL, sample=NULL, group=NULL,
-                      trend=FALSE, robust=TRUE)
+                      trend=FALSE, robust=TRUE, transform="asin")
 #    Testing for differences in cell type proportions
 #    Belinda Phipson
 #    29 July 2019
@@ -132,9 +137,11 @@ propeller <- function(x=NULL, clusters=NULL, sample=NULL, group=NULL,
         sample <- y$sample
         group <- y$group
     }
+    
+    if(is.null(transform)) transform <- "asin"
 
     # Get transformed proportions
-    prop.list <- getTransformedProps(clusters, sample)
+    prop.list <- getTransformedProps(clusters, sample, transform)
 
     # Calculate baseline proportions for each cluster
     baseline.props <- table(clusters)/sum(table(clusters))