+++ NEWS.md
@@ -1,3 +1,9 @@
+# speckle 0.0.2
+* Added functions to plot the mean variance relationship of the cell type
+counts and proportions
+* Added functions to estimate parameters of a Beta distribution 
+* Added logit transformation option to propeller
+
 # speckle 0.0.1
 
 * First version of the speckle package contains propeller functions to test for